+++ app/src/main/java/com/example/quizgame/MainActivity.java
@@ -4,7 +4,6 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import android.os.CountDownTimer;
-import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
@@ -19,16 +18,13 @@ public class MainActivity extends AppCompatActivity {
     Button opc1,opc2,opc3,opc4;
     TextView temporizador,puntaje,enunciado;
     Contador counter;
-    int seconds=60,minutes=1,ramdonselect=0,score=0,puntos=0,contador=0;
-    String idrandom="";
+    int seconds=60,minutes=1;
 
     private QuizLab QuizLab;
 
     private Usuario usuario;
     private Pregunta pregunta;
     private Respuesta respuesta;
-
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -46,267 +42,17 @@ public class MainActivity extends AppCompatActivity {
 
         QuizLab=QuizLab.get(this);
 
-        /*pregunta=new Pregunta();
-        pregunta.setPregunta("Cual es la flor nacional");
+        pregunta=new Pregunta();
+
+        /*pregunta.setPregunta("Cual es la flor nacional");
         pregunta.setValorPuntos(10);
         QuizLab.addPregunta(pregunta);*/
-
-        /*pregunta=new Pregunta();
-        pregunta.setPregunta("Eleija el nombre cientifico del Arbol Conacaste - Enterolobium cyclocarpum");
-        pregunta.setValorPuntos(20);
-        QuizLab.addPregunta(pregunta);*/
-
-        /*pregunta=new Pregunta();
-        pregunta.setPregunta("Eleija el nombre cientifico del Árbol siete camisas - Guapira witsbergeri");
-        pregunta.setValorPuntos(10);
-        QuizLab.addPregunta(pregunta);*/
-
-        /*pregunta=new Pregunta();
-        pregunta.setPregunta("Eleija el nombre cientifico Rana arborícola de ojos negros - Agalychnis moreletii");
-        pregunta.setValorPuntos(5);
-        QuizLab.addPregunta(pregunta);*/
-
-        /*pregunta=preguntas.get(0);
-        String q1=pregunta.getIdpregunta();
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(1);
-        respuesta.setRespuesta("Flor de Izote");
-        respuesta.setIdpregunta(q1);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Maquilishuat");
-        respuesta.setIdpregunta(q1);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Caña de Azúcar");
-        respuesta.setIdpregunta(q1);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Girasol");
-        respuesta.setIdpregunta(q1);
-        QuizLab.addRespuesta(respuesta);*/
-
-        /*pregunta=preguntas.get(1);
-        String q2=pregunta.getIdpregunta();
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Zea mays");
-        respuesta.setIdpregunta(q2);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(1);
-        respuesta.setRespuesta("Enterolobium cyclocarpum");
-        respuesta.setIdpregunta(q2);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Quercus");
-        respuesta.setIdpregunta(q2);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Tabebuia rosea");
-        respuesta.setIdpregunta(q2);
-        QuizLab.addRespuesta(respuesta);
-
-        pregunta=preguntas.get(2);
-        String q3 =pregunta.getIdpregunta();
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Phaseolus vulgaris");
-        respuesta.setIdpregunta(q3);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Spondias");
-        respuesta.setIdpregunta(q3);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Mangifera indica");
-        respuesta.setIdpregunta(q3);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(1);
-        respuesta.setRespuesta("Guapira witsbergeri");
-        respuesta.setIdpregunta(q3);
-        QuizLab.addRespuesta(respuesta);
-
-        pregunta=preguntas.get(3);
-        String q4 =pregunta.getIdpregunta();
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Persea americana");
-        respuesta.setIdpregunta(q4);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(1);
-        respuesta.setRespuesta("Agalychnis moreletii");
-        respuesta.setIdpregunta(q4);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Malus domestica");
-        respuesta.setIdpregunta(q4);
-        QuizLab.addRespuesta(respuesta);
-
-        respuesta=new Respuesta();
-        respuesta.setEstado(0);
-        respuesta.setRespuesta("Cucumis melo");
-        respuesta.setIdpregunta(q4);
-        QuizLab.addRespuesta(respuesta);*/
 
 
         List<Pregunta> preguntas=QuizLab.getPreguntas();
-        List<Respuesta> respuestas=QuizLab.getRespuestas();
-
         if (preguntas.size()>0){
-            ramdonselect=(int)(Math.random()*preguntas.size()+1);
-            while(ramdonselect>=preguntas.size()){
-                ramdonselect=(int)(Math.random());
-            }
-            pregunta= preguntas.get(ramdonselect);
+            pregunta= preguntas.get(0);
             enunciado.setText(pregunta.getPregunta());
-            idrandom=pregunta.getIdpregunta();
-
-            List<Respuesta> opciones=QuizLab.getOpciones(idrandom);
-            opc1.setText(opciones.get(0).getRespuesta());
-            opc2.setText(opciones.get(1).getRespuesta());
-            opc3.setText(opciones.get(2).getRespuesta());
-            opc4.setText(opciones.get(3).getRespuesta());
-
-            contador=1;
-
-            opc1.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-
-                    List<Respuesta> opciones=QuizLab.getOpciones(idrandom);
-                    puntos=CheckAnswer(opciones.get(0).getEstado());
-                    score=score+puntos;
-                    puntaje.setText("Puntaje: "+score+" pts");
-
-                    ramdonselect=(int)(Math.random()*preguntas.size()+1);
-                    while(ramdonselect>=preguntas.size()){
-                        ramdonselect=(int)(Math.random());
-                    }
-                    pregunta= preguntas.get(ramdonselect);
-                    enunciado.setText(pregunta.getPregunta());
-                    idrandom=pregunta.getIdpregunta();
-
-                    opciones=QuizLab.getOpciones(idrandom);
-                    opc1.setText(opciones.get(0).getRespuesta());
-                    opc2.setText(opciones.get(1).getRespuesta());
-                    opc3.setText(opciones.get(2).getRespuesta());
-                    opc4.setText(opciones.get(3).getRespuesta());
-
-                    contador ++;
-                }
-            });
-
-            opc2.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-
-                    List<Respuesta> opciones=QuizLab.getOpciones(idrandom);
-                    puntos=CheckAnswer(opciones.get(1).getEstado());
-                    score=score+puntos;
-                    puntaje.setText("Puntaje: "+score+" pts");
-
-                    ramdonselect=(int)(Math.random()*preguntas.size()+1);
-                    while(ramdonselect>=preguntas.size()){
-                        ramdonselect=(int)(Math.random());
-                    }
-                    pregunta= preguntas.get(ramdonselect);
-                    enunciado.setText(pregunta.getPregunta());
-                    idrandom=pregunta.getIdpregunta();
-
-                    opciones=QuizLab.getOpciones(idrandom);
-                    opc1.setText(opciones.get(0).getRespuesta());
-                    opc2.setText(opciones.get(1).getRespuesta());
-                    opc3.setText(opciones.get(2).getRespuesta());
-                    opc4.setText(opciones.get(3).getRespuesta());
-
-                    contador ++;
-                }
-            });
-
-            opc3.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-
-                    List<Respuesta> opciones=QuizLab.getOpciones(idrandom);
-                    puntos=CheckAnswer(opciones.get(2).getEstado());
-                    score=score+puntos;
-                    puntaje.setText("Puntaje: "+score+" pts");
-
-                    ramdonselect=(int)(Math.random()*preguntas.size()+1);
-                    while(ramdonselect>=preguntas.size()){
-                        ramdonselect=(int)(Math.random());
-                    }
-                    pregunta= preguntas.get(ramdonselect);
-                    enunciado.setText(pregunta.getPregunta());
-                    idrandom=pregunta.getIdpregunta();
-
-                    opciones=QuizLab.getOpciones(idrandom);
-                    opc1.setText(opciones.get(0).getRespuesta());
-                    opc2.setText(opciones.get(1).getRespuesta());
-                    opc3.setText(opciones.get(2).getRespuesta());
-                    opc4.setText(opciones.get(3).getRespuesta());
-
-                    contador ++;
-                }
-            });
-
-            opc4.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-
-                    List<Respuesta> opciones=QuizLab.getOpciones(idrandom);
-                    puntos=CheckAnswer(opciones.get(3).getEstado());
-                    score=score+puntos;
-                    puntaje.setText("Puntaje: "+score+" pts");
-
-                    ramdonselect=(int)(Math.random()*preguntas.size()+1);
-                    while(ramdonselect>=preguntas.size()){
-                        ramdonselect=(int)(Math.random());
-                    }
-                    pregunta= preguntas.get(ramdonselect);
-                    enunciado.setText(pregunta.getPregunta());
-                    idrandom=pregunta.getIdpregunta();
-
-                    opciones=QuizLab.getOpciones(idrandom);
-                    opc1.setText(opciones.get(0).getRespuesta());
-                    opc2.setText(opciones.get(1).getRespuesta());
-                    opc3.setText(opciones.get(2).getRespuesta());
-                    opc4.setText(opciones.get(3).getRespuesta());
-
-                    contador ++;
-                }
-            });
-
-
-
-
-
         }
         if(preguntas.size()<=0){
             enunciado.setText("Lo sentimos no se han encontrado preguntas en la base de datos");
@@ -316,22 +62,10 @@ public class MainActivity extends AppCompatActivity {
         counter = new Contador(120000,1000);
 
         counter.start();
-
     }
-    public int CheckAnswer(int estado){
-        int puntos=0;
 
-        if(estado ==1){
-            puntos=pregunta.getValorPuntos();
-        }
-        if (estado ==0){
-            puntos=0;
-        }
-        return puntos;
-    }//Función para validar respuesta
-
-    public void end(int contador){
-        temporizador.setText("Se acabó el tiempo\nPudiste responder "+contador+" preguntas.");
+    public void end(){
+        temporizador.setText("Se acabo el tiempo");
     }
 
 
@@ -360,7 +94,7 @@ public class MainActivity extends AppCompatActivity {
 
         @Override
         public void onFinish() {
-            end(contador);
+            end();
         }
 
         @Override
